+++ magick/signature.c
@@ -364,10 +364,10 @@ static void TransformSignature(SignatureInfo *signature_info)
 %
 %
 */
-static void UpdateSignature(SignatureInfo *signature_info,unsigned char *message,
-  size_t length)
+static void UpdateSignature(SignatureInfo *signature_info,
+  unsigned char *message,size_t length)
 {
-  register int
+  register size_t
     i;
 
   unsigned long
